02_centigrade_main/requirements.txt
pyserial
oauthlib
oauth2client
google-api-core
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth                       
google-auth-httplib2
googleapis-common-protos
gspread
virtualenv
numpy
matplotlib
cmake
dlib
face_recognition
opencv-python